worked for a few hours trying to get the windows build working
I was originally just going to commit these changes directly into the
source tree; however, I didn't quite want to do that.

There's some pretty gross nonsense that goes on to get >1.1 OpenGL
functionality working. Namely, you have to load every function you want
to use at runtime.

That doesn't seem to be the case on linux, as the practice is to link
against the driver, and just send it.

That's why this is going in its own branch.

Not to mention that I had to include a specific-ish build path. The env
vars that get used with gcc on win and linux are different and janky.

Uuuuugh