worked through some forking()
There was a problem with a deadlock in the original implementation,
and some colleagues and I have decided that the correct solution
is to have a parent write thread and a parent read thread from
the child.

Now the only thing left to do is work out how to elegantly have
threading in the program.

The correct answer is probably making a sys_* abstraction.